feat: allow accessing exported Rust functions in JavaScript
