+++ site/content/_index.md
@@ -67,6 +67,7 @@ cargo install --locked wasm-bindgen-cli
 ```
 
 ## App Setup
+
 Get setup with your favorite `wasm-bindgen` based framework. [Yew](https://github.com/yewstack/yew) & [Seed](https://github.com/seed-rs/seed) are the most popular options today, but there are others. Trunk will work with any `wasm-bindgen` based framework. The easiest way to ensure that your application launches properly is to [setup your app as an executable](https://doc.rust-lang.org/cargo/guide/project-layout.html) with a standard `main` function:
 
 ```rust
@@ -85,7 +86,7 @@ Trunk uses a source HTML file to drive all asset building and bundling. Trunk al
 </html>
 ```
 
-`trunk build` will produce the following HTML at `dist/index.html`, along with the compiled scss, WASM & the JS loader for the WASM:
+`trunk-ng build` will produce the following HTML at `dist/index.html`, along with the compiled scss, WASM & the JS loader for the WASM:
 
 ```html
 <html>
@@ -96,7 +97,8 @@ Trunk uses a source HTML file to drive all asset building and bundling. Trunk al
   </head>
   <body>
     <script type="module">
-      import init from '/index-7eeee8fa37b7636a.js';
+      import init, * as bindings from '/index-7eeee8fa37b7636a.js';
+      window.wasmBindings = bindings;
       init('/index-7eeee8fa37b7636a_bg.wasm');
     </script>
   </body>
@@ -105,20 +107,29 @@ Trunk uses a source HTML file to drive all asset building and bundling. Trunk al
 
 The contents of your `dist` dir are now ready to be served on the web.
 
+## JavaScript interoperability
+
+Trunk will create the necessary JavaScript code to bootstrap and run the WebAssembly based application. It will also
+include all JavaScript snippets generated by `wasm-bindgen` for interfacing with JavaScript functionality.
+
+By default, functions exported from Rust, using `wasm-bingen`, can be accessed in the JavaScript code through the global
+variable `window.wasmBindings`. This behavior can be disabled, and the name can be customized. 
+
 # Next Steps
+
 That's not all! Trunk has even more useful features. Head on over to the following sections to learn more about how to use Trunk effectively.
 
 - [Assets](@/assets.md): learn about all of Trunk's supported asset types.
 - [Configuration](@/configuration.md): learn about Trunk's configuration system and how to use the Trunk proxy.
 - [Commands](@/commands.md): learn about Trunk's CLI commands for use in your development workflows.
-- Join us on Discord by following this link <a href="https://discord.gg/JEPdBujTDr"><img src="https://img.shields.io/discord/793890238267260958?logo=discord&style=flat-square" style="vertical-align:text-top;" alt="Discord Chat"/></a>
+- Join us on Discord by following this link [![](https://img.shields.io/discord/793890238267260958?logo=discord&style=flat-square "Discord Chat")](https://discord.gg/JEPdBujTDr)
 
 # Contributing
+
 Anyone and everyone is welcome to contribute! Please review the [CONTRIBUTING.md](https://github.com/thedodd/trunk/blob/master/CONTRIBUTING.md) document for more details. The best way to get started is to find an open issue, and then start hacking on implementing it. Letting other folks know that you are working on it, and sharing progress is a great approach. Open pull requests early and often, and please use GitHub's draft pull request feature.
 
 # License
-<p>
-    <span><img src="https://img.shields.io/badge/license-MIT%2FApache--2.0-blue?style=flat-square" alt="license badge"/></span>
-    <br/>
-    trunk is licensed under the terms of the MIT License or the Apache License 2.0, at your choosing.
-</p>
+
+![](https://img.shields.io/badge/license-MIT%2FApache--2.0-blue?style=flat-square "license badge")
+
+trunk-ng (as well as trunk) is licensed under the terms of the MIT License or the Apache License 2.0, at your choosing.
